refactor(plugin): extract runtime_context.rs (#1079)
Moves the 8 host_rcv_* callbacks + HOST_RUNTIME_CONTEXT_VTABLE
static + host_runtime_context_vtable() accessor + the two test
modules (null-handle guards + tier-1 wire-format) out of
host_services/mod.rs into a sibling submodule. mod.rs re-exports
HOST_RUNTIME_CONTEXT_VTABLE and host_runtime_context_vtable so
external callers (crate::core::context::runtime_context's shim
constructors) compile unchanged.

cargo test -p streamlib-engine --lib unchanged at 1407 passed / 2
pre-existing Deno failures / 126 ignored.
